# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ID8 Operator API
  version: 0.1.0
  description: |
    MVP API contract for orchestrating prompt-to-production runs with HITL gates.
servers:
  - url: https://api.id8.local
paths:
...
  /v1/projects/{projectId}/design/generate:
    post:
      summary: Generate or regenerate design artifacts
      operationId: generateDesign
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DesignGenerateRequest'
      responses:
        '202':
          description: Design job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactResponse'
  /v1/projects/{projectId}/design/feedback:
    post:
      summary: Submit targeted design feedback for regeneration
      operationId: submitDesignFeedback
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DesignFeedbackRequest'
      responses:
        '202':
          description: Feedback accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactResponse'
...
components:
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
  schemas:
    DesignProvider:
      type: string
      enum:
        - stitch_mcp
        - internal_spec
        - manual_upload
    ModelProfile:
      type: string
      enum:
        - primary
        - customtools
        - fallback
...
    ArtifactType:
      type: string
      enum:
        - prd
        - design_spec
        - tech_plan
        - code_snapshot
        - security_report
        - deploy_report
...
    StitchAuthMethod:
      type: string
      enum:
        - api_key
        - oauth_access_token
    StitchAuthPayload:
      type: object
      properties:
        auth_method:
          $ref: '#/components/schemas/StitchAuthMethod'
        api_key:
          type: string
        oauth_token:
          type: string
        goog_user_project:
          type: string
...
    DesignGenerateRequest:
      type: object
      properties:
        provider:
          $ref: '#/components/schemas/DesignProvider'
        model_profile:
          $ref: '#/components/schemas/ModelProfile'
        prompt_constraints:
          type: object
          additionalProperties: true
        stitch_auth:
          $ref: '#/components/schemas/StitchAuthPayload'
      required:
        - provider
        - model_profile
    DesignFeedbackRequest:
      type: object
      properties:
        target_screen_id:
          type: string
        target_component_id:
          type: string
        feedback_text:
          type: string
        stitch_auth:
          $ref: '#/components/schemas/StitchAuthPayload'
      required:
        - feedback_text
...
    ProjectArtifact:
      type: object
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        run_id:
          type: string
          format: uuid
        artifact_type:
          $ref: '#/components/schemas/ArtifactType'
        version:
          type: integer
        content:
          type: object
          additionalProperties: true
        model_profile:
          $ref: '#/components/schemas/ModelProfile'
        created_at:
          type: string
          format: date-time
      required:
        - id
        - project_id
        - run_id
        - artifact_type
        - version
        - content
        - created_at
    ArtifactResponse:
      type: object
      properties:
        artifact:
          $ref: '#/components/schemas/ProjectArtifact'
      required:
        - artifact
...